code-review/requirements.txt
python-gitlab==6.3.0
PyGitHub==2.8.1
llm==0.27.1
llm-openrouter==0.5
